fix: remove unused include wait.h
We moved our internal 'wait.h' to 'common/wait.h' but forgot the include
in 'lttng-events.c'. This was a stale include as we don't use anything
from wait.h here. On Linux this went unnoticed as there is a system
'wait.h' which is an alias for 'sys/wait.h'. However, there is no such
system header on FreeBSD and it results in a build failure.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I10034b65e4cacae0ef729ce07d505a37f0b2f1cc