Remove -Wsystem-headers from tests
The objective of add -Wsystem-headers to the tests Makefiles was to
catch warnings in the tracepoint headers when they are installed in the
compiler system header path. However the tests are always built against
the headers inside the tree which are not in the system header path.

A side effect is that it causes problems on some system where the system
headers generate warnings and thus prevent us from building with
warnings treated as errors.

Change-Id: I65529c50cf48e49ba34692e6d501e3218be8d41c
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>