Make futex compat internal to liblttng-ust
This compat header originated in userspace RCU where it's used across
multiple shared objects hence the need to have the mutexes as public
weak symbols, in our case it's only used internally by liblttng-ust so
we can hide them.

If we end up using this compat header in another library in this project
we will have to use the same scheme, but in the meantime, don't expose
those symbols in the ABI.

Move the code to the liblttng-ust directory to make sure it's not used
by other libraries.

Change-Id: Ibfaef9448eeaf6247b42f71d6b6d8de234d79a3c
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>