Fix: Java application context: pass application context argument to callbacks
The callbacks require the complete application context field names to
compare against the tables received from the TLS area.

We also need to free memory allocated by for the context event field and
context name in a destroy callback.

Therefore, allocate a data structure for each application context, and
use that data structure as private data. This allows callback to reach
the context name, and a destroy callback to free the allocated memory.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ia34fe00f5945fb58ef991f9d40dec36ac9ab322c