tracepoints: increase dlopen failure message level from debug to critical
Print the failure message associated with failing to find
lttng-ust-tracepoint.so as a "Critical: " message, because when this
situation occurs, it indeeds makes part of that application's
instrumentation invisible to the tracer.

Similarly to debug message, this critical message is only shown if
LTTNG_UST_DEBUG is defined for the compile unit or if the
LTTNG_UST_DEBUG environment variable is set.

In addition however, if LTTNG_UST_ABORT_ON_CRITICAL is defined at
compile-time, or if the application is run with the
LTTNG_UST_ABORT_ON_CRITICAL environment variable set, the construction
will call abort() on failure to find lttng-ust-tracepoint.so.

This should make it easier for end-users to identify deployment issues
which prevent the lttng-ust tracer from being aware of application
tracepoints.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I2ddcfd593eae699a2c18ef85049ac2239dd41411