Properly capture java variables at configure
Some Java related variables aren't currently properly captured when
running for example:

    ./configure JAVACFLAGS="-g"

The JAVACFLAGS variable will be used during the configure phase but
won't be when running make afterwards.

This is due to the AX_JAVA_OPTIONS macro which adds non-standard
configure switches like '--with-javac-flags' which as a side-effect
inhibit the capture of the variable.

Remove this macro and use AC_ARG_VAR to capture these variables and
bring the Java support in line with C/C++.

Change-Id: I9159e34c1c30a10309200bfc8e0fd067000d14ce
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>